backend/requirements.txt
# Core API framework
fastapi
uvicorn[standard]

# Environment variable loading
python-dotenv

# Groq SDK (LLaMA 3.3 70B — free & fast)
# FIX: Replaced google-generativeai (wrong) with groq (actually used in main.py)
# Get your free API key at: https://console.groq.com
groq

# File upload support
python-multipart

# Data validation
pydantic

# PDF extraction (text-based PDFs) — Strategy 1
pymupdf

# Scanned PDF / image OCR — Strategy 2
# Requires Tesseract binary installed separately:
#   Windows: https://github.com/UB-Mannheim/tesseract/wiki
#   Mac:     brew install tesseract
#   Ubuntu:  sudo apt install tesseract-ocr poppler-utils
pytesseract
pdf2image
Pillow

# Excel and CSV parsing — Strategy 3
pandas
openpyxl

# PDF report generation (used in /api/generate-report endpoint)
# FIX: Changed version from 2.8.7 (does not exist) to 2.7.9 (latest stable)
fpdf2